Extract LogLevel
