+++ Adjust/adjust/src/main/java/com/adjust/sdk/ActivityHandler.java
@@ -345,7 +345,7 @@ public class ActivityHandler extends HandlerThread implements IActivityHandler {
         deviceInfo = new DeviceInfo(adjustConfig.context, adjustConfig.sdkPrefix);
 
         if (adjustConfig.environment == AdjustConfig.ENVIRONMENT_PRODUCTION) {
-            logger.setLogLevel(Logger.LogLevel.ASSERT);
+            logger.setLogLevel(LogLevel.ASSERT);
         } else {
             logger.setLogLevel(adjustConfig.logLevel);
         }
