+++ Adjust/adjust/src/main/java/com/adjust/sdk/Logger.java
@@ -3,20 +3,6 @@ package com.adjust.sdk;
 import android.util.Log;
 
 public interface Logger {
-
-    public enum LogLevel {
-        VERBOSE(Log.VERBOSE), DEBUG(Log.DEBUG), INFO(Log.INFO), WARN(Log.WARN), ERROR(Log.ERROR), ASSERT(Log.ASSERT);
-        final int androidLogLevel;
-
-        LogLevel(final int androidLogLevel) {
-            this.androidLogLevel = androidLogLevel;
-        }
-
-        public int getAndroidLogLevel() {
-            return androidLogLevel;
-        }
-    }
-
     public void setLogLevel(LogLevel logLevel);
 
     public void setLogLevelString(String logLevelString);
